requirements.txt
langgraph==0.2.74
langsmith
langchain-openai
langchain_community
langchain_experimental
python-dotenv
ipython
tavily-python
pyodbc
websockets
azure-search-documents
azure-identity